Changes in functions and layout
In this change, the player class starting stats have been changed to placeholder text, not quite sure if this is a good solution to when the party levels up the stats reset. We will see how this works out. Added a few other counters and moved the primary buttons closer to the top. I moved the about text into its own page for a better experience. Almost ready to do css work.